refactor(components): renamed component file names to match up with class names
